refactor: rename library/ to board-set/, drop redundant prefix
Folder name matches contents — every file was board-set-* — and the
prefix no longer doubles up.
