+++ src/features/board/board-set/info-dialog.tsx
